Add wearing pressure suit
r6453
2021-11-13

## ISSUE
1. https://github.com/mars-sim/mars-sim/issues/502

## NEW
1. Add wearStandardClothing() and unwearStandardClothing
   in Person.

## CHANGE
1. Wear pressure suit inside vehicle.
2. Rename "counter pressure suit" to "pressure suit".
3. Implement ItemHolder in EVASuit.
4. Modify createEquipment() to store a pressure suit inside
   the newly created EVA suit.
5. Modify in InventoryTabPanel.
   - equipmentTable.addMouseListener()
   - loadModel()

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/equipment/Equipment.java
@@ -248,7 +248,7 @@ public abstract class Equipment extends Unit implements Indoor, Salvagable {
 	 * @param resource
 	 * @return quantity
 	 */
-	public double getItemResourceStored(int resource) {
+	public int getItemResourceStored(int resource) {
 		return microInventory.getItemResourceStored(resource);
 	}
